% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toweRdefense-package.R
\docType{package}
\name{toweRdefense-package}
\alias{toweRdefense}
\alias{toweRdefense-package}
\title{toweRdefense: Toer Defense game in R}
\description{
Play a tower defense game in R.
}
\author{
\strong{Maintainer}: Alex Axthelm \email{nonogramr@AlexAxthelm.com} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]

}
\keyword{internal}
